config-shepherd/requirements.txt
PyYAML>=6.0.1
jsonschema>=4.21.0
pytest>=8.0.0
pytest-cov>=4.1.0
